fix: sql-duplicate-class-fqcn
