+++ extensions/store/sql/federated-catalog-cache-sql/src/main/java/org/eclipse/edc/catalog/cache/sql/SqlFederatedCatalogCacheExtension.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2024 Amadeus IT Group
+ *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,15 +8,16 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Amadeus IT Group - initial API and implementation
+ *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
  *
  */
 
-package org.eclipse.edc.catalog.store.sql;
+package org.eclipse.edc.catalog.cache.sql;
 
-import org.eclipse.edc.catalog.store.sql.schema.postgres.PostgresDialectStatements;
-import org.eclipse.edc.crawler.spi.TargetNode;
-import org.eclipse.edc.crawler.spi.TargetNodeDirectory;
+import org.eclipse.edc.catalog.cache.sql.schema.postgres.PostgresDialectStatements;
+import org.eclipse.edc.catalog.spi.FederatedCatalogCache;
+import org.eclipse.edc.connector.controlplane.catalog.spi.Catalog;
+import org.eclipse.edc.connector.controlplane.catalog.spi.Dataset;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.runtime.metamodel.annotation.Provides;
@@ -29,11 +30,11 @@ import org.eclipse.edc.sql.bootstrapper.SqlSchemaBootstrapper;
 import org.eclipse.edc.transaction.datasource.spi.DataSourceRegistry;
 import org.eclipse.edc.transaction.spi.TransactionContext;
 
-@Provides(TargetNodeDirectory.class)
-@Extension(value = "SQL target node directory")
-public class SqlTargetNodeDirectoryExtension implements ServiceExtension {
+@Provides(FederatedCatalogCache.class)
+@Extension(value = "SQL federated catalog cache")
+public class SqlFederatedCatalogCacheExtension implements ServiceExtension {
 
-    @Setting(description = "The datasource to be used", defaultValue = DataSourceRegistry.DEFAULT_DATASOURCE, key = "edc.sql.store.targetnodedirectory.datasource")
+    @Setting(description = "The datasource to be used", defaultValue = DataSourceRegistry.DEFAULT_DATASOURCE, key = "edc.sql.store.federatedcatalog.datasource")
     private String dataSourceName;
 
     @Inject
@@ -41,7 +42,7 @@ public class SqlTargetNodeDirectoryExtension implements ServiceExtension {
     @Inject
     private TransactionContext trxContext;
     @Inject(required = false)
-    private TargetNodeStatements statements;
+    private FederatedCatalogCacheStatements statements;
     @Inject
     private TypeManager typeManager;
 
@@ -53,17 +54,17 @@ public class SqlTargetNodeDirectoryExtension implements ServiceExtension {
 
     @Override
     public void initialize(ServiceExtensionContext context) {
-        typeManager.registerTypes(TargetNode.class);
-        var targetNodeDirectory = new SqlTargetNodeDirectory(dataSourceRegistry, dataSourceName, trxContext,
+        typeManager.registerTypes(Catalog.class, Dataset.class);
+        var store = new SqlFederatedCatalogCache(dataSourceRegistry, dataSourceName, trxContext,
                 typeManager.getMapper(), queryExecutor, getStatementImpl());
-        context.registerService(TargetNodeDirectory.class, targetNodeDirectory);
-        sqlSchemaBootstrapper.addStatementFromResource(dataSourceName, "target-node-directory-schema.sql");
+        context.registerService(FederatedCatalogCache.class, store);
+        sqlSchemaBootstrapper.addStatementFromResource(dataSourceName, "cache-schema.sql");
     }
 
     /**
      * returns an externally-provided sql statement dialect, or postgres as a default
      */
-    private TargetNodeStatements getStatementImpl() {
+    private FederatedCatalogCacheStatements getStatementImpl() {
         return statements != null ? statements : new PostgresDialectStatements();
     }
