+++ extensions/store/sql/federated-catalog-cache-sql/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -12,4 +12,4 @@
 #
 #
 
-org.eclipse.edc.catalog.store.sql.SqlFederatedCatalogCacheExtension
+org.eclipse.edc.catalog.cache.sql.SqlFederatedCatalogCacheExtension
