+++ extensions/store/sql/target-node-directory-sql/src/main/java/org/eclipse/edc/catalog/directory/sql/SqlTargetNodeDirectory.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.catalog.store.sql;
+package org.eclipse.edc.catalog.directory.sql;
 
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
