+++ extensions/store/sql/target-node-directory-sql/src/main/java/org/eclipse/edc/catalog/directory/sql/SqlTargetNodeDirectoryExtension.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2024 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *  Copyright (c) 2024 Amadeus IT Group
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,16 +8,15 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *       Amadeus IT Group - initial API and implementation
  *
  */
 
-package org.eclipse.edc.catalog.store.sql;
+package org.eclipse.edc.catalog.directory.sql;
 
-import org.eclipse.edc.catalog.spi.FederatedCatalogCache;
-import org.eclipse.edc.catalog.store.sql.schema.postgres.PostgresDialectStatements;
-import org.eclipse.edc.connector.controlplane.catalog.spi.Catalog;
-import org.eclipse.edc.connector.controlplane.catalog.spi.Dataset;
+import org.eclipse.edc.catalog.directory.sql.schema.postgres.PostgresDialectStatements;
+import org.eclipse.edc.crawler.spi.TargetNode;
+import org.eclipse.edc.crawler.spi.TargetNodeDirectory;
 import org.eclipse.edc.runtime.metamodel.annotation.Extension;
 import org.eclipse.edc.runtime.metamodel.annotation.Inject;
 import org.eclipse.edc.runtime.metamodel.annotation.Provides;
@@ -30,11 +29,11 @@ import org.eclipse.edc.sql.bootstrapper.SqlSchemaBootstrapper;
 import org.eclipse.edc.transaction.datasource.spi.DataSourceRegistry;
 import org.eclipse.edc.transaction.spi.TransactionContext;
 
-@Provides(FederatedCatalogCache.class)
-@Extension(value = "SQL federated catalog cache")
-public class SqlFederatedCatalogCacheExtension implements ServiceExtension {
+@Provides(TargetNodeDirectory.class)
+@Extension(value = "SQL target node directory")
+public class SqlTargetNodeDirectoryExtension implements ServiceExtension {
 
-    @Setting(description = "The datasource to be used", defaultValue = DataSourceRegistry.DEFAULT_DATASOURCE, key = "edc.sql.store.federatedcatalog.datasource")
+    @Setting(description = "The datasource to be used", defaultValue = DataSourceRegistry.DEFAULT_DATASOURCE, key = "edc.sql.store.targetnodedirectory.datasource")
     private String dataSourceName;
 
     @Inject
@@ -42,7 +41,7 @@ public class SqlFederatedCatalogCacheExtension implements ServiceExtension {
     @Inject
     private TransactionContext trxContext;
     @Inject(required = false)
-    private FederatedCatalogCacheStatements statements;
+    private TargetNodeStatements statements;
     @Inject
     private TypeManager typeManager;
 
@@ -54,17 +53,17 @@ public class SqlFederatedCatalogCacheExtension implements ServiceExtension {
 
     @Override
     public void initialize(ServiceExtensionContext context) {
-        typeManager.registerTypes(Catalog.class, Dataset.class);
-        var store = new SqlFederatedCatalogCache(dataSourceRegistry, dataSourceName, trxContext,
+        typeManager.registerTypes(TargetNode.class);
+        var targetNodeDirectory = new SqlTargetNodeDirectory(dataSourceRegistry, dataSourceName, trxContext,
                 typeManager.getMapper(), queryExecutor, getStatementImpl());
-        context.registerService(FederatedCatalogCache.class, store);
-        sqlSchemaBootstrapper.addStatementFromResource(dataSourceName, "cache-schema.sql");
+        context.registerService(TargetNodeDirectory.class, targetNodeDirectory);
+        sqlSchemaBootstrapper.addStatementFromResource(dataSourceName, "target-node-directory-schema.sql");
     }
 
     /**
      * returns an externally-provided sql statement dialect, or postgres as a default
      */
-    private FederatedCatalogCacheStatements getStatementImpl() {
+    private TargetNodeStatements getStatementImpl() {
         return statements != null ? statements : new PostgresDialectStatements();
     }
