+++ extensions/store/sql/target-node-directory-sql/src/main/java/org/eclipse/edc/catalog/directory/sql/TargetNodeStatements.java
@@ -12,7 +12,7 @@
  *
  */
 
-package org.eclipse.edc.catalog.store.sql;
+package org.eclipse.edc.catalog.directory.sql;
 
 import org.eclipse.edc.spi.query.QuerySpec;
 import org.eclipse.edc.sql.statement.SqlStatements;
