+++ extensions/store/sql/target-node-directory-sql/src/main/resources/META-INF/services/org.eclipse.edc.spi.system.ServiceExtension
@@ -12,4 +12,4 @@
 #
 #
 
-org.eclipse.edc.catalog.store.sql.SqlTargetNodeDirectoryExtension
+org.eclipse.edc.catalog.directory.sql.SqlTargetNodeDirectoryExtension
