+++ extensions/store/sql/target-node-directory-sql/src/test/java/org/eclipse/edc/catalog/directory/sql/SqlTargetNodeDirectoryExtensionTest.java
@@ -1,5 +1,5 @@
 /*
- *  Copyright (c) 2023 Bayerische Motoren Werke Aktiengesellschaft (BMW AG)
+ *  Copyright (c) 2024 Amadeus IT Group
  *
  *  This program and the accompanying materials are made available under the
  *  terms of the Apache License, Version 2.0 which is available at
@@ -8,16 +8,15 @@
  *  SPDX-License-Identifier: Apache-2.0
  *
  *  Contributors:
- *       Bayerische Motoren Werke Aktiengesellschaft (BMW AG) - initial API and implementation
+ *       Amadeus IT Group - initial API and implementation
  *
  */
 
-package org.eclipse.edc.catalog.store.sql;
+package org.eclipse.edc.catalog.directory.sql;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import org.eclipse.edc.catalog.spi.FederatedCatalogCache;
-import org.eclipse.edc.connector.controlplane.catalog.spi.Catalog;
-import org.eclipse.edc.connector.controlplane.catalog.spi.Dataset;
+import org.eclipse.edc.crawler.spi.TargetNode;
+import org.eclipse.edc.crawler.spi.TargetNodeDirectory;
 import org.eclipse.edc.junit.extensions.DependencyInjectionExtension;
 import org.eclipse.edc.spi.system.ServiceExtensionContext;
 import org.eclipse.edc.spi.system.configuration.Config;
@@ -33,7 +32,7 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 @ExtendWith(DependencyInjectionExtension.class)
-public class SqlFederatedCatalogCacheExtensionTest {
+public class SqlTargetNodeDirectoryExtensionTest {
 
     private final TypeManager typeManager = mock();
 
@@ -44,16 +43,16 @@ public class SqlFederatedCatalogCacheExtensionTest {
     }
 
     @Test
-    void shouldInitializeTheStore(SqlFederatedCatalogCacheExtension extension, ServiceExtensionContext context) {
+    void shouldInitializeTheStore(SqlTargetNodeDirectoryExtension extension, ServiceExtensionContext context) {
         var config = mock(Config.class);
         when(context.getConfig()).thenReturn(config);
         when(config.getString(any(), any())).thenReturn("test");
 
         extension.initialize(context);
 
-        var service = context.getService(FederatedCatalogCache.class);
-        assertThat(service).isInstanceOf(SqlFederatedCatalogCache.class);
+        var service = context.getService(TargetNodeDirectory.class);
+        assertThat(service).isInstanceOf(SqlTargetNodeDirectory.class);
 
-        verify(typeManager).registerTypes(Catalog.class, Dataset.class);
+        verify(typeManager).registerTypes(TargetNode.class);
     }
 }
